+++ src/RazorLight/IEngineHandler.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Threading.Tasks;
 using RazorLight.Caching;
 using RazorLight.Compilation;
+using System;
 
 namespace RazorLight
 {
@@ -21,7 +22,9 @@ namespace RazorLight
 		Task<ITemplatePage> CompileTemplateAsync(string key);
 
 		Task<string> CompileRenderAsync<T>(string key, T model, ExpandoObject? viewBag = null);
+		Task<string> CompileRenderAsync(string key, object? model, Type modelType, ExpandoObject? viewBag = null);
 		Task<string> CompileRenderStringAsync<T>(string key, string content, T model, ExpandoObject? viewBag = null);
+		Task<string> CompileRenderStringAsync(string key, string content, object? model, Type modelType, ExpandoObject? viewBag = null);
 
 		Task<string> RenderTemplateAsync<T>(ITemplatePage templatePage, T model, ExpandoObject? viewBag = null);
 		Task RenderTemplateAsync<T>(ITemplatePage templatePage, T model, TextWriter textWriter, ExpandoObject? viewBag = null);